Changed: Extract TableSelection class.
